Extract BitcoinConnectorMock into test module
Also, rename `/tests/bitcoin.rs` to
`/tests/bitcoin_transaction_pattern_e2e.rs`, since it holds a very
specific test and there are/will be more than one test file related to
Bitcoin.
